# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoChat 聊天/好友/房间管理后端接口",
        "title": "GoChat Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/tokens/refresh": {
            "get": {
                "description": "使用刷新令牌 Cookie 刷新访问令牌，并重新设置刷新令牌 Cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "刷新访问令牌",
                "responses": {
                    "200": {
                        "description": "刷新成功，返回新的访问令牌",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_authorization_port_http.RefreshAccessTokenResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "刷新令牌无效或已过期",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
        "/chats/private-messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的私聊消息列表，可基于 base_id 游标和 limit 分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取私聊消息列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页大小，默认20，最大100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回私聊消息列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_chat_port_http.ListPrivateMessagesResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向指定用户发送一条私聊消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "发送私聊消息",
                "parameters": [
                    {
                        "description": "发送私聊消息请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_chat_port_http.SendPrivateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
        "/chats/private-messages/read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向指定房间发送一条消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "发送房间消息",
                "parameters": [
                    {
                        "description": "发送房间消息请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_chat_port_http.ReadPrivateMessagesRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
        "/chats/rooms/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户所在房间的消息列表，可基于 base_id 游标和 limit 分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "获取房间消息列表",
                "parameters": [
                    {
                        "enum": [
                            ""
                        ],
                        "type": "string",
                        "description": "分页游标，返回该ID之前的消息",
                        "name": "base_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页大小，默认20，最大100",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回房间消息列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/gochat_internal_shared_api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_chat_port_http.ListRoomMessagesResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向指定房间发送一条消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "发送房间消息",
                "parameters": [
                    {
                        "description": "发送房间消息请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_chat_port_http.SendRoomMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/gochat_internal_shared_api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gochat_internal_chat_domain.MessageState": {
            "type": "string",
            "enum": [
                "undelivered",
                "delivered",
                "read"
            ],
            "x-enum-varnames": [
                "MessageStateUndelivered",
                "MessageStateDelivered",
                "MessageStateRead"
            ]
        },
        "gochat_internal_chat_dto.PrivateMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recipient_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                },
                "sender_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                },
                "sent_at": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/gochat_internal_chat_domain.MessageState"
                }
            }
        },
        "gochat_internal_chat_dto.RoomMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.MessageID"
                },
                "room_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.RoomID"
                },
                "sender_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                },
                "sent_at": {
                    "type": "string"
                },
                "states": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/gochat_internal_chat_domain.MessageState"
                    }
                }
            }
        },
...
        "gochat_internal_shared_api.Code": {
            "type": "integer",
            "enum": [
                200,
                500,
                400,
                401,
                402,
                403,
                404
            ],
            "x-enum-varnames": [
                "CodeSuccess",
                "CodeServerError",
                "CodeTimeout",
                "CodeInvalidParam",
                "CodeInvalidToken",
                "CodeNotFound",
                "CodeServiceUnavailable"
            ]
        },
        "gochat_internal_shared_api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/gochat_internal_shared_api.Code"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "gochat_internal_shared_kernel.MessageID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
...
        "gochat_internal_shared_kernel.RoomID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
        "gochat_internal_shared_kernel.UserID": {
            "type": "string",
            "enum": [
                ""
            ],
            "x-enum-varnames": [
                "EmptyUserID"
            ]
        },
...
        "internal_authorization_port_http.RefreshAccessTokenResponseData": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
...
        "internal_chat_port_http.ListPrivateMessagesResponseData": {
            "type": "object",
            "properties": {
                "private_messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gochat_internal_chat_dto.PrivateMessage"
                    }
                }
            }
        },
        "internal_chat_port_http.ListRoomMessagesResponseData": {
            "type": "object",
            "properties": {
                "room_messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gochat_internal_chat_dto.RoomMessage"
                    }
                }
            }
        },
        "internal_chat_port_http.ReadPrivateMessagesRequest": {
            "type": "object",
            "required": [
                "sender_id"
            ],
            "properties": {
                "sender_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                }
            }
        },
...
        "internal_chat_port_http.SendPrivateMessageRequest": {
            "type": "object",
            "required": [
                "content",
                "recipient_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 1000
                },
                "recipient_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.UserID"
                }
            }
        },
        "internal_chat_port_http.SendRoomMessageRequest": {
            "type": "object",
            "required": [
                "content",
                "room_id"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 1000
                },
                "room_id": {
                    "$ref": "#/definitions/gochat_internal_shared_kernel.RoomID"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}